+++ packages/stream_feeds/lib/src/models/feed_suggestion_data.dart
@@ -4,6 +4,7 @@ import 'package:stream_core/stream_core.dart';
 import '../generated/api/models.dart';
 import 'feed_data.dart';
 import 'feed_id.dart';
+import 'feed_input_data.dart';
 import 'feed_member_data.dart';
 import 'follow_data.dart';
 import 'user_data.dart';
@@ -75,13 +76,13 @@ extension FeedSuggestionResponseMapper on FeedSuggestionResponse {
         ),
         memberCount: memberCount,
         name: name,
-        ownCapabilities: [...?ownCapabilities],
+        ownCapabilities: ownCapabilities ?? const [],
         ownMembership: ownMembership?.toModel(),
         ownFollowings: ownFollowings?.map((f) => f.toModel()).toList(),
         ownFollows: ownFollows?.map((f) => f.toModel()).toList(),
         pinCount: pinCount,
         updatedAt: updatedAt,
-        visibility: visibility,
+        visibility: visibility?.let(FeedVisibility.new),
         custom: custom,
       ),
       algorithmScores: algorithmScores,