+++ packages/stream_feeds_test/lib/src/helpers/test_data.dart
@@ -220,7 +220,7 @@ PollResponseData createDefaultPollResponse({
     voteCountsByOption: latestVotesByOption.map(
       (k, e) => MapEntry(k, e.length),
     ),
-    votingVisibility: 'visibility',
+    votingVisibility: PollResponseDataVotingVisibility.public,
     options: options,
   );
 }